test(mpsc): keep buffer reclamation checks inline
